refactor: extract Squid laser engine into its own module
Moves the ~660 lines of laser-engine code (status types, parser/builder,
SquidLaserEngineBase, SquidLaserEngine, SquidLaserEngine_Simulation,
SquidLaserEngineError, public LASER_CHANNEL_ORDER) out of
control/serial_peripherals.py and into a dedicated
control/squid_laser_engine.py.

Updates the three importers (control/microscope.py, control/laser_engine_widget.py,
tests/control/test_squid_laser_engine.py) to pull from the new module.
serial_peripherals.py loses its now-unused IntEnum/dataclass/crc32/QObject/Signal
/Iterable/List imports.

Also reverts control/_def.py to ship-default values (USE_SQUID_LASER_ENGINE=False,
SQUID_LASER_ENGINE_SN=None) — the prior commit accidentally captured a local
hardware-testing edit. Local users continue to flip these to True+real-SN as
needed; we just don't commit those changes.
